refactors openseadragon, #1215
